+++ src/systems.ts
@@ -41,8 +41,8 @@ export function renderingSystem(engine: Engine) {// some total of all, uses laye
             else if (shape.shape instanceof Circle) {
                 drawFn = () => circleRenderingSystem(engine, id);
             }
-            else if(shape.shape instanceof Triangle){
-                drawFn = () => triangleRenderingSystem(engine,id);
+            else if (shape.shape instanceof Triangle) {
+                drawFn = () => triangleRenderingSystem(engine, id);
             }
             if (drawFn) {
                 drawCalls.push({
@@ -81,11 +81,17 @@ function rectangleRenderingSystem(engine: Engine, id: EntityId) {
     const shape = em.getComponent(id, Shape);
     if (!transform) throw new Error("TRANSFORM NULL IN RECTANGLE RENDERING STSTEM !" + id);
     if (!shape) throw new Error("SHAPE NULL IN RECTANGLE RENDERING SYSTEM! " + id);
-    if ((shape.shape instanceof Rectangle)) {
+    const rectangle = shape.shape;
+    if ((rectangle instanceof Rectangle)) {
         const x: number = transform.globalPosition.position.x;
         const y: number = transform.globalPosition.position.y;
-        const w: number = shape.shape.width;
-        const h: number = shape.shape.height;
+        const w: number = rectangle.width;
+        const h: number = rectangle.height;
+        const rotationInRadians = (Math.PI * rectangle.rotation) / 180
+        const offset: Vector2 = {
+            x: rectangle.centered ? -w / 2 : 0,
+            y: rectangle.centered ? -h / 2 : 0
+        }
 
         ctx.save();
 
@@ -95,7 +101,13 @@ function rectangleRenderingSystem(engine: Engine, id: EntityId) {
 
         ctx.beginPath();
 
-        ctx.rect(x, y, w, h);
+        ctx.translate(x, y);
+
+        ctx.rotate(rotationInRadians);
+
+        ctx.scale(transform.scale.value.x,transform.scale.value.y);
+
+        ctx.rect(offset.x, offset.y, w, h);
 
         ctx.fill();
 
@@ -109,7 +121,7 @@ function rectangleRenderingSystem(engine: Engine, id: EntityId) {
 
             ctx.beginPath();
 
-            ctx.rect(x, y, w, h);
+            ctx.rect(offset.x, offset.y, w, h);
 
             ctx.stroke();
 
@@ -146,7 +158,11 @@ function circleRenderingSystem(engine: Engine, id: EntityId) {
 
         ctx.beginPath();
 
-        ctx.arc(x, y, r, 0, Math.PI * 2);
+        ctx.translate(x,y);
+
+        ctx.scale(transform.scale.value.x,transform.scale.value.y);
+
+        ctx.arc(0, 0, r, 0, Math.PI * 2);
 
         ctx.fill();
 
@@ -160,7 +176,7 @@ function circleRenderingSystem(engine: Engine, id: EntityId) {
 
             ctx.beginPath();
 
-            ctx.arc(x, y, r, 0, Math.PI * 2);
+            ctx.arc(0, 0, r, 0, Math.PI * 2);
 
             ctx.stroke();
 
@@ -181,52 +197,64 @@ function triangleRenderingSystem(engine: Engine, id: EntityId) {
     const ctx = engine.getCtx();
     const transform = em.getComponent(id, Transform);
     const shape = em.getComponent(id, Shape);
-    if (!transform) throw new Error("TRANSFORM NULL IN RECTANGLE RENDERING STSTEM !" + id);
-    if (!shape) throw new Error("SHAPE NULL IN RECTANGLE RENDERING SYSTEM! " + id);
-    if ((shape.shape instanceof Triangle)) {
-        const s1: Vector2 = shape.shape.s1;
-        const s2: Vector2 = shape.shape.s2;
-        const s3: Vector2 = shape.shape.s3;
+    if (!transform) throw new Error("TRANSFORM NULL IN TRIANGLE RENDERING SYSTEM !" + id);
+    if (!shape) throw new Error("SHAPE NULL IN TRIANGLE RENDERING SYSTEM! " + id);
+    const triangle = shape.shape;
+    if (!(triangle instanceof Triangle)) {
+        console.warn("SHAPE IS NOT A TRIANGLE BUT IS IN TRIANGLE RENDERING SYSTEM! " + id);
+        return;
+    }
 
-        ctx.save();
+    const pos = transform.globalPosition.position;
+    const rotationInRadians = (Math.PI * triangle.rotation) / 180;
 
-        //SHAPE
+    // Step 1: calculate local triangle points relative to center or p1
+    let localP1: Vector2, localP2: Vector2, localP3: Vector2;
 
-        ctx.fillStyle = shape.color;
-        ctx.globalAlpha = shape.alpha;
+    if (triangle.centered) {
+        const centroid = {
+            x: (triangle.p1.x + triangle.p2.x + triangle.p3.x) / 3,
+            y: (triangle.p1.y + triangle.p2.y + triangle.p3.y) / 3,
+        };
+
+        localP1 = { x: triangle.p1.x - centroid.x, y: triangle.p1.y - centroid.y };
+        localP2 = { x: triangle.p2.x - centroid.x, y: triangle.p2.y - centroid.y };
+        localP3 = { x: triangle.p3.x - centroid.x, y: triangle.p3.y - centroid.y };
+    } else {
+        // treat p1 as origin
+        localP1 = { x: 0, y: 0 };
+        localP2 = { x: triangle.p2.x - triangle.p1.x, y: triangle.p2.y - triangle.p1.y };
+        localP3 = { x: triangle.p3.x - triangle.p1.x, y: triangle.p3.y - triangle.p1.y };
+    }
+
+    // Step 2: draw using local coordinates, then translate/rotate
+    ctx.save();
+
+    ctx.fillStyle = shape.color;
+    ctx.globalAlpha = shape.alpha;
+
+    ctx.translate(pos.x, pos.y);
+    ctx.rotate(rotationInRadians);
+    ctx.scale(transform.scale.value.x,transform.scale.value.y);
+
+    ctx.beginPath();
+    ctx.moveTo(localP1.x, localP1.y);
+    ctx.lineTo(localP2.x, localP2.y);
+    ctx.lineTo(localP3.x, localP3.y);
+    ctx.closePath();
+    ctx.fill();
+
+    if (shape.outlineEnabled) {
+        ctx.strokeStyle = shape.outlineColor;
+        ctx.lineWidth = shape.outlineWidth;
 
         ctx.beginPath();
-
-        ctx.moveTo(s1.x, s1.y);
-        ctx.lineTo(s1.x, s1.y);
-        ctx.lineTo(s2.x, s2.y);
-        ctx.lineTo(s3.x, s3.y);
-
-        ctx.fill();
-
+        ctx.moveTo(localP1.x, localP1.y);
+        ctx.lineTo(localP2.x, localP2.y);
+        ctx.lineTo(localP3.x, localP3.y);
         ctx.closePath();
-
-        //OUTLINE
-
-        if (shape.outlineEnabled) {
-            ctx.strokeStyle = shape.outlineColor;
-            ctx.lineWidth = shape.outlineWidth;
-
-            ctx.beginPath();
-
-            ctx.moveTo(s1.x, s1.y);
-            ctx.lineTo(s1.x, s1.y);
-            ctx.lineTo(s2.x, s2.y);
-            ctx.lineTo(s3.x, s3.y);
-
-            ctx.stroke();
-
-            ctx.closePath();
-        }
-
-        ctx.restore();
+        ctx.stroke();
     }
-    else {
-        console.warn("SHAPE IS NOT A TRIANGLE BUT IS IN RECTANGLE RENDERING SYSTEM! " + id);
-    }
+
+    ctx.restore();
 };